Adopt exact measurements in `setout` and `joiner`
Replace the signed local `setout::Length` with the shared strictly-positive `Length` and signed `Offset` domains. Add named relations for adjusting sizes, translating coordinates, and locating offsets by positive distances, including full-range reverse solving.

Make `setout_joiner` the explicit exact-to-floating boundary, migrate the basilica roof setout, and bump the canonical fingerprint schema to version 2 with a documented rebuild migration.

Use exact `Length` and `Angle` values for authored `joiner` rejection limits and contact thresholds while retaining explicitly named floating seams for analytic geometry-derived values.

Close `em-u8k7` with no-std, workspace, documentation, and byte-identical basilica export validation.

+++ crates/setout/src/evaluation.rs
@@ -2072,7 +2072,7 @@ fn encode_exactness(exactness: &ExactnessTrace, encoder: &mut CanonicalEncoder) 
         ExactnessTrace::ImportedFloat(quantization) => {
             encoder.u8(3);
             encoder.u64(quantization.source_bits);
-            encoder.i64(quantization.selected_iota);
+            encoder.i128(quantization.selected_iota);
             encoder.u64(quantization.error_iota_bits);
         }
     }
